refactor(sim): move single-caller helpers to their owning sub-files
The first-pass split left the "Internal helpers" block on the root
`impl Simulation` in sim.rs, even though each private helper had
exactly one calling sub-file. Reviewer flagged the coupling leftover;
moving each helper next to its caller finishes the decomposition.

* `register_hall_call_for_rider` → `sim/rider.rs` (only caller:
  `spawn_rider_inner` in same file). Stays `fn` since it's now
  co-located with its caller.
* `ensure_hall_call` → `sim/calls.rs` as `pub(super) fn` — still
  called from `sim/rider.rs::register_hall_call_for_rider`, so
  pub(super) is required.
* `ack_latency_for`, `ack_latency_for_stop`, `ack_latency_for_car`,
  `ensure_car_call` → `sim/calls.rs` as `fn` — all internal helpers
  of `calls.rs` only.
* `brake_target_stop` (free fn) → `sim/destinations.rs` — only caller
  is `abort_movement` in the same file. `super::brake_target_stop`
  call in `abort_movement` becomes unqualified `brake_target_stop`.

sim.rs drops 586 → 397 lines. All 604 lib tests + 156 doc tests pass;
`api_surface_tests` unchanged. No public-API change.
