+++ crates/elevator-core/src/sim/destinations.rs
@@ -9,6 +9,7 @@ use crate::entity::{ElevatorId, EntityId};
 use crate::error::SimError;
 use crate::events::Event;
 use crate::stop::StopRef;
+use crate::world::World;
 
 impl super::Simulation {
     // ── Destination queue (imperative dispatch) ────────────────────
@@ -157,7 +158,7 @@ impl super::Simulation {
             return Ok(());
         };
 
-        let Some(brake_stop) = super::brake_target_stop(&self.world, pos, vel, brake_pos) else {
+        let Some(brake_stop) = brake_target_stop(&self.world, pos, vel, brake_pos) else {
             return Ok(());
         };
 
@@ -189,4 +190,48 @@ impl super::Simulation {
         }
         Ok(())
     }
+}
+
+/// Pick the stop to park at when aborting an in-flight movement.
+///
+/// Tries three strategies in order:
+///
+/// 1. **Closest stop at or past `brake_pos` in the direction of travel.**
+///    The car can decelerate into it naturally without overshoot.
+/// 2. **Farthest stop still in the direction of travel (end-of-line).**
+///    If the car is too close to the end of the line to fit a full
+///    deceleration, pick the terminal stop ahead of it; the movement
+///    system's overshoot-snap will absorb the small residual distance.
+/// 3. **Nearest stop overall.** Only reachable when the car has no
+///    stops ahead of it at all (e.g., single-stop worlds or the car is
+///    already past the final stop); also handles the `vel == 0` case.
+///
+/// Returns `None` only if the world has no stops at all.
+fn brake_target_stop(world: &World, pos: f64, vel: f64, brake_pos: f64) -> Option<EntityId> {
+    let dir = vel.signum();
+    if dir != 0.0 {
+        let ahead_of_brake = world
+            .iter_stops()
+            .filter(|(_, stop)| (stop.position() - brake_pos) * dir >= 0.0)
+            .min_by(|(_, a), (_, b)| {
+                (a.position() - pos)
+                    .abs()
+                    .total_cmp(&(b.position() - pos).abs())
+            })
+            .map(|(id, _)| id);
+        if ahead_of_brake.is_some() {
+            return ahead_of_brake;
+        }
+        let ahead_of_car = world
+            .iter_stops()
+            .filter(|(_, stop)| (stop.position() - pos) * dir >= 0.0)
+            .max_by(|(_, a), (_, b)| {
+                ((a.position() - pos) * dir).total_cmp(&((b.position() - pos) * dir))
+            })
+            .map(|(id, _)| id);
+        if ahead_of_car.is_some() {
+            return ahead_of_car;
+        }
+    }
+    world.find_nearest_stop(brake_pos)
 }
